move old stuff to legacy
