+++ packages/webview_flutter/webview_flutter_web/example/integration_test/legacy/webview_flutter_test.dart
@@ -9,7 +9,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter/widgets.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:integration_test/integration_test.dart';
-import 'package:webview_flutter_web_example/web_view.dart';
+import 'package:webview_flutter_web_example/legacy/web_view.dart';
 
 void main() {
   IntegrationTestWidgetsFlutterBinding.ensureInitialized();
