+++ packages/webview_flutter/webview_flutter_web/example/lib/main.dart
@@ -9,7 +9,7 @@ import 'dart:typed_data';
 import 'package:flutter/material.dart';
 import 'package:webview_flutter_platform_interface/webview_flutter_platform_interface_legacy.dart';
 
-import 'web_view.dart';
+import 'legacy/web_view.dart';
 
 void main() {
   runApp(const MaterialApp(home: _WebViewExample()));
