+++ packages/webview_flutter/webview_flutter_web/lib/src/webview_flutter_web_legacy.dart
@@ -10,7 +10,7 @@ import 'package:flutter/gestures.dart';
 import 'package:flutter/widgets.dart';
 import 'package:flutter_web_plugins/flutter_web_plugins.dart';
 import 'package:webview_flutter_platform_interface/webview_flutter_platform_interface_legacy.dart';
-import 'shims/dart_ui.dart' as ui;
+import 'legacy/shims/dart_ui.dart' as ui;
 
 /// Builds an iframe based WebView.
 ///
