+++ packages/webview_flutter/webview_flutter_web/test/legacy/webview_flutter_web_test.mocks.dart
@@ -1,5 +1,5 @@
 // Mocks generated by Mockito 5.3.2 from annotations
-// in webview_flutter_web/test/webview_flutter_web_test.dart.
+// in webview_flutter_web/test/legacy/webview_flutter_web_test.dart.
 // Do not manually edit this file.
 
 // ignore_for_file: no_leading_underscores_for_library_prefixes
@@ -15,7 +15,8 @@ import 'package:webview_flutter_platform_interface/src/legacy/types/types.dart'
     as _i8;
 import 'package:webview_flutter_platform_interface/webview_flutter_platform_interface_legacy.dart'
     as _i9;
-import 'package:webview_flutter_web/webview_flutter_web.dart' as _i10;
+import 'package:webview_flutter_web/src/webview_flutter_web_legacy.dart'
+    as _i10;
 
 // ignore_for_file: type=lint
 // ignore_for_file: avoid_redundant_argument_values
